initial check-in
example of config and rpm .spec for building a stand-alone xlator
outside the glusterfs buildl tree.

See the INSTALL file for (terse) instructions for building.

For the sake of the example I have copied the experimental/posix2
xlator from the glusterfs master branch.

Note that this does not address the issue of including the xlator
in the xlator graph when you create a volume from the gluster CLI.

Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>